Finished up POST and Get for Conductores Profile Pics
Finished up methods to POST a picture for a user (can work for update if
FrontEnd does some workarounds), one can retrieve as well the picture
from a person with a GET Method. All of these methods are protected by a
CSRF Token. Production server working with the new implementations.