Changed Naming Method for Profile Pic
Now the names for each picture are linked directly to the user_id, thanks to that frontend doesn't need any sort of workaround to replace profile pictures. They will always will stay the same, the POST method to upload a picture can now be used to also update a picture.