Done LoginTest using PHPUnit
As a security measure, use composer install after pulling. Made a Unit Test for the Login endpoint, made a good request and then did a bad request. Both tests work. In order to test with VS Code easily, download "PHPUnit Test Explorer". In order for the tests to run smoothly make a copy of .env with the name .env.testing, when running tests PHPUnit will override this config file instead of the .env. once that's done, be careful that on this .env.testing file, on the field that says APP_URL the url is pointing towards the environment you plan to test on (It can be localhost, or production server).