Refactored order of functions, finished CRUD
Finished Update, Show and destroy functions for VehicleController all the DAO and BL classes have been implemented successfully in order to make CRUD operations with Vehicle.
Also reorganized the order of the functions in both Controller, BL and DAO so that it can be easily read, it follows the order of "CRUD" from top to bottom.