Finished connection methods for Vehiculo and Conductor
Filled up logic in ConductorController in order to extend the finding of a Vehicle by Id for a Conductor and also to get a List of all the Vehicles from a Conductor. 
VehiculoBL and VehiculoDAO have also been loaded with the functions to retrieve data with this logic and report in case of errors. 
Added the base model for a Conductor_VehiculoController, since it is needed in order to register when a Vehicle is assigned to a Conductor.