Finished up Report and Check mechanism for Conductores
Finished all the requirements that involve backend for US - 9, made a new Controller and respective BL/DAO classes in order to save reports for an specific Conductor, along with a timestamp, created Reporte migration and model as well.
Refactored on the fly a method in order to be reusable for other methods in ConductorBL and ConductorDAO.
Added route in order to make reports.
Made a small change to the json response whenever a Punto de control is moved to the next one, now the json response has the Next Punto de Control for the driver.